requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg==3.0.10
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
